Optimise rendering, fix zoom and add colours
Rendering has been optimised so that the program can run at around
40FPS while zoomed in 2x. It still gets very laggy as you zoom out and
more is rendered.

Colours have been added to distinguish nodes from segments, and more
will be added for further distinction between types of roads, etc.

Zoom has been fixed so that both nodes and segments scale with the zoom
without being randomly cropped out.

+++ assignment-1/src/comp261/assignment1/helper/LocationHelper.java
@@ -15,8 +15,8 @@ public class LocationHelper {
 	 * @param lat - Latitude to convert
 	 * @return X location
 	 */
-	public static double getLocX(double lat) {
-		return (lat - Graph.CENTRE_LAT) * Graph.SCALE_LAT * Graph.zoom;
+	public static float getLocX(double lat) {
+		return (float) ((lat - Graph.CENTRE_LAT) * Graph.SCALE_LAT * Graph.zoom);
 	}
 
 	/**
@@ -26,7 +26,7 @@ public class LocationHelper {
 	 * @param lon - Longitude to convert
 	 * @return Y location
 	 */
-	public static double getLocY(double lon) {
-		return (lon - Graph.CENTRE_LON) * Graph.SCALE_LON * Graph.zoom;
+	public static float getLocY(double lon) {
+		return (float) ((lon - Graph.CENTRE_LON) * Graph.SCALE_LON * Graph.zoom);
 	}
 }
